refactor: split stats renderer into modules
